refactor: 기존 mail 로직 mail 모듈로 이동 (#117)
