+++ core/core-api/src/main/java/cmc/mellyserver/common/event/NotificationEventHandler.java
@@ -1,11 +1,12 @@
 package cmc.mellyserver.common.event;
 
 
+import static cmc.mellyserver.notification.constants.NotificationConstants.COMMENT_LIKE_NOTI_CONTENT;
+
 import cmc.mellyserver.dbcore.notification.enums.NotificationType;
 import cmc.mellyserver.domain.comment.event.CommentEnrollEvent;
 import cmc.mellyserver.domain.comment.event.CommentLikeEvent;
-import cmc.mellyserver.domain.notification.NotificationService;
-import cmc.mellyserver.message.MessageService;
+import cmc.mellyserver.notification.NotificationService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Async;
@@ -13,17 +14,15 @@ import org.springframework.stereotype.Component;
 import org.springframework.transaction.event.TransactionPhase;
 import org.springframework.transaction.event.TransactionalEventListener;
 
-import static cmc.mellyserver.message.constants.NotificationConstants.COMMENT_LIKE_NOTI_CONTENT;
-
 
 @Slf4j
 @Component
 @RequiredArgsConstructor
 public class NotificationEventHandler {
 
-    private final MessageService pushService;
+    private final NotificationService pushService;
 
-    private final NotificationService notificationService;
+    private final cmc.mellyserver.domain.notification.NotificationService notificationService;
 
 
     @Async
