--- infra/mail/src/main/java/cmc/mellyserver/mail/MailApplication.java
@@ -1,13 +1,0 @@
-package cmc.mellyserver.mail;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MailApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(MailApplication.class, args);
-    }
-
-}
