+++ infra/notification/src/main/java/cmc/mellyserver/notification/constants/NotificationConstants.java
@@ -1,4 +1,4 @@
-package cmc.mellyserver.message.constants;
+package cmc.mellyserver.notification.constants;
 
 public abstract class NotificationConstants {
